Ajuste no servidor
