+++ package.json
@@ -15,7 +15,7 @@
   "dependencies": {
     "axios": "^0.21.0",
     "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
+    "dotenv": "^8.6.0",
     "express": "^4.17.1",
     "helmet": "^8.0.0",
     "morgan": "^1.10.0",
